Create new class for utterance/category display
